Refactor : token 관련 폴더 #1
