--- user-module/src/main/java/com/example/user_module/auth/service/RefreshTokenService.java
@@ -1,7 +1,0 @@
-package com.example.user_module.auth.service;
-
-import com.example.user_module.common.security.jwt.RefreshTokenResponseDTO;
-
-public interface RefreshTokenService {
-    RefreshTokenResponseDTO refreshToken(String refreshToken);
-}
