+++ user-module/src/main/java/com/example/user_module/common/security/jwt/dto/RefreshRes.java
@@ -1,4 +1,4 @@
-package com.example.user_module.common.security.jwt;
+package com.example.user_module.common.security.jwt.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RefreshTokenResponseDTO {
+public class RefreshRes {
     private String accessToken;
     private String refreshToken;
 }
