+++ user-module/src/main/java/com/example/user_module/common/security/jwt/service/RefreshTokenService.java
@@ -1,0 +1,7 @@
+package com.example.user_module.common.security.jwt.service;
+
+import com.example.user_module.common.security.jwt.dto.RefreshRes;
+
+public interface RefreshTokenService {
+    RefreshRes refreshToken(String refreshToken);
+}
